# Ignore non-semantic changes to objects