#### What type of PR is this?

/kind bug

#### What this PR does / why we need it:

Ignores changes that only affect the timestamp in the managed fields. That can happen when someone merely updates a Quantity from 1024 to 1Ki or when accidentally removing a value that was a default. This change will reduce the total number of writes to etcd for usages which apply frequent changes to configuration with no semantic value.

#### Which issue(s) this PR fixes:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #95460

#### Does this PR introduce a user-facing change?

```release-note
NONE
```