@sttts 2022-01-20
Sgtm. My concerns about semantics of the edge cases with nil and empty lists/maps are resolved with the latest updates. 

There have been questions about performance before when having this in every update code-path. Has those been addressed or have there been any benchmarks to mitigate those concerns?