@alexzielenski 2022-02-15
I've ran this change, and the original code, quite a few times against the a WIP field validation benchmark `put-ignore`.

The results were surprising, indicating a possible small performance improvement.

```
name                                               old time/op    new time/op    delta
FieldValidation/Put/put-ignore-validation-16         2.32ms ± 8%    2.24ms ± 9%  -3.16%  (p=0.000 n=806+874)
FieldValidation/Put/put-ignore-validation-yaml-16    2.75ms ±11%    2.67ms ±12%  -2.85%  (p=0.000 n=742+734)
```