@alexzielenski 2022-03-01
/cc @liggitt 

Took another benchmark with the new comparison taken in isolation.

The source code for this benchmark is here: https://gist.github.com/alexzielenski/8715898d08976e93eb59a42bb1d9f7b5

There are two benchmarks: for each of structured types and unstructured types. For each benchmark, a shallow object (v1.Endpoints) was tested (generated using fuzzer), and then a deep object (v1.Pod).

```
BenchmarkIgnoreManagedFieldsTimestampTransformerStructured
BenchmarkIgnoreManagedFieldsTimestampTransformerStructured/shallow
BenchmarkIgnoreManagedFieldsTimestampTransformerStructured/shallow-8         	    8595	    151961 ns/op	   32891 B/op	    1499 allocs/op
BenchmarkIgnoreManagedFieldsTimestampTransformerStructured/deep
BenchmarkIgnoreManagedFieldsTimestampTransformerStructured/deep-8            	     235	   5390087 ns/op	 1253875 B/op	   33811 allocs/op
```

```
BenchmarkIgnoreManagedFieldsTimestampTransformerUntructured
BenchmarkIgnoreManagedFieldsTimestampTransformerUntructured/shallow
BenchmarkIgnoreManagedFieldsTimestampTransformerUntructured/shallow-8         	      46	  25047611 ns/op	 5193434 B/op	  188879 allocs/op
BenchmarkIgnoreManagedFieldsTimestampTransformerUntructured/deep
BenchmarkIgnoreManagedFieldsTimestampTransformerUntructured/deep-8            	     129	   8737373 ns/op	 1669834 B/op	   58405 allocs/op
```

I expect that the shallow object was slower for unstructured types due to the excessive amount of reflection done to iterate the map keys (very wide fan out). The deep object does not fan out nearly as much.