@liggitt 2022-03-04
> There are two benchmarks: for each of structured types and unstructured types. For each benchmark, a shallow object (v1.Endpoints) was tested (generated using fuzzer), and then a deep object (v1.Pod).

huh... those are significantly more expensive than I hoped