@alexzielenski 2022-03-10
> > There are two benchmarks: for each of structured types and unstructured types. For each benchmark, a shallow object (v1.Endpoints) was tested (generated using fuzzer), and then a deep object (v1.Pod).
> 
> huh... those are significantly more expensive than I hoped

I added a log to the benchmark to get a sense of what objects sort are being measured and ran again. The log prints the size of the object formatted as JSON in bytes.

```
    /Users/zielenski/go/src/k8s.io/kubernetes/test/integration/apiserver/timestamp_transformer_test.go:52: Shallow (Endpoints): 19839 bytes
    /Users/zielenski/go/src/k8s.io/kubernetes/test/integration/apiserver/timestamp_transformer_test.go:55: Deep (Pod): 170521 bytes
BenchmarkIgnoreManagedFieldsTimestampTransformerStructured/shallow
BenchmarkIgnoreManagedFieldsTimestampTransformerStructured/shallow-8         	    8953	    133800 ns/op	   45755 B/op	     999 allocs/op
BenchmarkIgnoreManagedFieldsTimestampTransformerStructured/deep
BenchmarkIgnoreManagedFieldsTimestampTransformerStructured/deep-8            	     618	   1868402 ns/op	  530219 B/op	   11178 allocs/op
```

This ran faster than the last time I ran the benchmark (I wonder if my computer was under load). The 2ms score for the `deep` benchmark is on a 170kb as json. Hopefully this provides some more context.

Here are results of the unstructured benchmarks along with the input size:

```
    /Users/zielenski/go/src/k8s.io/kubernetes/test/integration/apiserver/timestamp_transformer_test.go:116: Shallow (Endpoints): 1725411 bytes
    /Users/zielenski/go/src/k8s.io/kubernetes/test/integration/apiserver/timestamp_transformer_test.go:119: Deep (Pod): 610822 bytes
BenchmarkIgnoreManagedFieldsTimestampTransformerUntructured/shallow
BenchmarkIgnoreManagedFieldsTimestampTransformerUntructured/shallow-8         	      48	  25991266 ns/op	 5193476 B/op	  188879 allocs/op
BenchmarkIgnoreManagedFieldsTimestampTransformerUntructured/deep
BenchmarkIgnoreManagedFieldsTimestampTransformerUntructured/deep-8            	     122	  10176093 ns/op	 1826131 B/op	   63847 allocs/op
```

This also provides possibly a hint to the question of why the "shallow" unstructured benchmark results  seemed inexplicably faster than "deep": my "shallow" fuzzed object for this unstructured benchmark is not small (>1.5MB!). Also keep in mind that these "unstructured" tests populate most fields of the object, rather than a small subset.