@apelisse 2022-03-10
OK So here's how I read this number, transforming them in human-readable:
- 133 µs for a 20k endpoint
- 1.9 ms for a 170k pod
- 26 ms for a 1.8M endpoint,
- 10ms for a 610k pods

I don't think that's that slow after-all?