@liggitt 2022-03-21
as a floor to add to every request, it seems high. that's also an unfortunate number of allocs... I didn't realize a readonly deepequal check involved so many