@lavalamp 2022-03-30
What are we blocked on here? I consider the bug this is fixing pretty critical.

(I also consider that number of allocs to be slightly ridiculous for comparison checks, but I haven't looked at the code here)