@liggitt 2022-03-30
> What are we blocked on here? I consider the bug this is fixing pretty critical.

uncertainty whether that's a reasonable number of allocs / compute to add to every write request