@liggitt 2022-03-30
> @liggitt what are numbers you could live with?

I don't know what our numbers are currently at, but the allocs especially seem egregious.

Did we get scaleability runs with/without this? We're actually in the middle of recovering from a scale regression today (https://github.com/kubernetes/kubernetes/issues/109146), so we'd need clean signal before getting data on this.

If this merges and we discover it causes problems, what's our recourse? Wait for a fix/backport/patch release?