@alexzielenski 2022-03-30
@liggitt @lavalamp You guys are correct to be suspicious of the number of allocations. I attached a debugger to the benchmark and I have found that the branch the test is taking the "transformation" code path, rather than simply the "comparison" code path:

https://github.com/kubernetes/kubernetes/pull/106388/files#diff-8604303e065a13f40ea019ec0b9a2fb733f44b5c6fbdd5b4b9238b442e94f527R58

This could very likely explain the high number of deep copies.

This is surprising since the benchmark tests two identical objects, [one being a deep copy of the other](https://github.com/kubernetes/kubernetes/pull/106388/files#diff-3e2d935ca4990d4678efcf482d4de1f9fe647109f846b9663fab39a7ae38d5fbR111). 

I'll post back in a bit with either updated benchmark results if it is a problem with the benchmark, or more interesting results if there is a bug with the deepCopy/comparison.