@apelisse 2022-04-05
I think that's clearly an improvement and I don't think, even for controller regularly updating the status, that this would fall in the worst case scenario?